Chore: Migrate repo to TypeScript
