--- .eslintrc.cjs
@@ -1,5 +1,0 @@
-'use strict'
-
-module.exports = {
-  extends: ['eloquence/node'],
-}
