+++ .eslintrc.js
@@ -1,0 +1,21 @@
+'use strict'
+
+module.exports = {
+  root: true,
+
+  extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
+  ignorePatterns: ['node_modules', 'examples'],
+  reportUnusedDisableDirectives: true,
+
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    ecmaVersion: 'latest',
+    ecmaFeatures: {},
+    tsconfigRootDir: __dirname,
+    project: ['./tsconfig.json'],
+  },
+
+  plugins: ['@typescript-eslint'],
+
+  rules: {},
+}
