+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
-  // Disable TS and implicit type checking for examples screenshots
-  "js/ts.implicitProjectConfig.checkJs": false,
-  "typescript.validate.enable": false
+  "editor.codeActionsOnSave": {
+    "source.organizeImports": true
+  }
 }
