+++ package.json
@@ -2,7 +2,6 @@
   "name": "radical-vscode",
   "displayName": "Radical",
   "description": "A dark theme for radical hacking inspired by retro futuristic design",
-  "type": "module",
   "version": "3.2.1",
   "publisher": "dhedgecock",
   "license": "ISC",
@@ -35,14 +34,19 @@
     ]
   },
   "scripts": {
-    "build": "node build.mjs",
-    "format": "prettier --write *.md build.mjs theme/**.mjs",
-    "lint": "NODE_ENV=test eslint build.mjs theme/**.mjs"
+    "build": "ts-node ./theme/build.ts",
+    "format": "prettier --write *.md theme",
+    "lint": "NODE_ENV=test eslint theme"
   },
   "devDependencies": {
+    "@types/chroma-js": "2.1.4",
+    "@types/node": "18.11.17",
+    "@typescript-eslint/eslint-plugin": "5.47.0",
+    "@typescript-eslint/parser": "5.47.0",
     "chroma-js": "2.4.2",
-    "eslint-config-eloquence": "21.2.0",
+    "eslint": "8.30.0",
     "prettier": "2.8.1",
+    "ts-node": "10.9.1",
     "typescript": "4.9.4"
   }
 }
